# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Grupo {
  id: number;
  nome: string;
  dataCriacao: string;
  totalParticipantes: number;
}

export interface Participante {
  id: number;
  nome: string;
  grupoId: number;
}

export interface GrupoDetalhe {
  id: number;
  nome: string;
  dataCriacao: string;
  participantes: Participante[];
}

export interface GrupoInput {
  /** @minLength 1 */
  nome: string;
  /** Optional list of participant names to add on creation */
  participantes?: string[];
}

export interface GrupoUpdate {
  /** @minLength 1 */
  nome?: string;
}

export interface Transferencia {
  pagadorId: number;
  pagadorNome: string;
  recebedorId: number;
  recebedorNome: string;
  valor: number;
}

export type DespesaCategoria = typeof DespesaCategoria[keyof typeof DespesaCategoria];


export const DespesaCategoria = {
  restaurante: 'restaurante',
  transporte: 'transporte',
  hospedagem: 'hospedagem',
  lazer: 'lazer',
  compras: 'compras',
  outros: 'outros',
} as const;

export interface Despesa {
  id: number;
  descricao: string;
  valor: number;
  pagadorId: number;
  pagadorNome: string;
  grupoId: number;
  participantes: Participante[];
  dataCriacao: string;
  categoria: DespesaCategoria;
}

export interface GrupoResumo {
  grupoId: number;
  nome: string;
  saldoTotal: number;
  aReceber: number;
  aPagar: number;
  transferencias: Transferencia[];
  despesasRecentes: Despesa[];
}

export interface ParticipanteInput {
  /** @minLength 1 */
  nome: string;
}

export type DespesaInputCategoria = typeof DespesaInputCategoria[keyof typeof DespesaInputCategoria];


export const DespesaInputCategoria = {
  restaurante: 'restaurante',
  transporte: 'transporte',
  hospedagem: 'hospedagem',
  lazer: 'lazer',
  compras: 'compras',
  outros: 'outros',
} as const;

export interface DespesaInput {
  /** @minLength 1 */
  descricao: string;
  /** @minimum 0.01 */
  valor: number;
  pagadorId: number;
  /** @minItems 1 */
  participanteIds: number[];
  categoria?: DespesaInputCategoria;
}

export type PagamentoStatus = typeof PagamentoStatus[keyof typeof PagamentoStatus];


export const PagamentoStatus = {
  pendente: 'pendente',
  pago: 'pago',
} as const;

export interface Pagamento {
  id: number;
  pagadorId: number;
  pagadorNome: string;
  recebedorId: number;
  recebedorNome: string;
  valor: number;
  status: PagamentoStatus;
  dataCriacao: string;
}

export interface DividasResumo {
  grupoId: number;
  nome: string;
  transferencias: Transferencia[];
  pagamentos: Pagamento[];
}

export interface PagamentoInput {
  pagadorId: number;
  recebedorId: number;
  /** @minimum 0.01 */
  valor: number;
}