# lib/api-zod/src/generated/types/despesa.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */
import type { DespesaCategoria } from './despesaCategoria';
import type { Participante } from './participante';

export interface Despesa {
  id: number;
  descricao: string;
  valor: number;
  pagadorId: number;
  pagadorNome: string;
  grupoId: number;
  participantes: Participante[];
  dataCriacao: string;
  categoria: DespesaCategoria;
}