# lib/api-zod/src/generated/types/despesaInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */
import type { DespesaInputCategoria } from './despesaInputCategoria';

export interface DespesaInput {
  /** @minLength 1 */
  descricao: string;
  /** @minimum 0.01 */
  valor: number;
  pagadorId: number;
  /** @minItems 1 */
  participanteIds: number[];
  categoria?: DespesaInputCategoria;
}