# lib/api-zod/src/generated/types/dividasResumo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */
import type { Pagamento } from './pagamento';
import type { Transferencia } from './transferencia';

export interface DividasResumo {
  grupoId: number;
  nome: string;
  transferencias: Transferencia[];
  pagamentos: Pagamento[];
}