# lib/api-zod/src/generated/types/grupoDetalhe.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */
import type { Participante } from './participante';

export interface GrupoDetalhe {
  id: number;
  nome: string;
  dataCriacao: string;
  participantes: Participante[];
}