# lib/api-zod/src/generated/types/grupoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */

export interface GrupoInput {
  /** @minLength 1 */
  nome: string;
  /** Optional list of participant names to add on creation */
  participantes?: string[];
}