# lib/api-zod/src/generated/types/grupoResumo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */
import type { Despesa } from './despesa';
import type { Transferencia } from './transferencia';

export interface GrupoResumo {
  grupoId: number;
  nome: string;
  saldoTotal: number;
  aReceber: number;
  aPagar: number;
  transferencias: Transferencia[];
  despesasRecentes: Despesa[];
}