# lib/api-zod/src/generated/types/pagamentoInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */

export interface PagamentoInput {
  pagadorId: number;
  recebedorId: number;
  /** @minimum 0.01 */
  valor: number;
}