# lib/api-zod/src/generated/types/transferencia.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Split – API for shared expense management
 * OpenAPI spec version: 0.1.0
 */

export interface Transferencia {
  pagadorId: number;
  pagadorNome: string;
  recebedorId: number;
  recebedorNome: string;
  valor: number;
}